flatten mod structure
