--- crates/tfw/src/data/mod.rs
@@ -1,7 +1,0 @@
-mod sysparams;
-mod systems;
-
-pub mod prelude {
-    pub use super::sysparams::*;
-    pub use super::systems::*;
-}
