+++ crates/tfw/src/scope.rs
