--- crates/tfw/tests/app/screen/empty/mod.rs
@@ -1,10 +1,0 @@
-use crate::prelude::*;
-
-mod screen;
-pub mod prelude {
-    pub use super::screen::EmptyScreen;
-}
-
-pub fn plugin(app: &mut App) {
-    app.add_plugins(screen::plugin);
-}
