--- crates/tfw/tests/app/screen/lifecycle/mod.rs
@@ -1,1 +1,0 @@
-mod screen;
