--- crates/tfw/tests/app/screen/load_strategy/mod.rs
@@ -1,1 +1,0 @@
-mod screen;
